+++ src/api/src/lib.rs
@@ -12,7 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-mod column_def;
 pub mod error;
 pub mod helper;
 pub mod prometheus;
