+++ src/common/grpc-expr/Cargo.toml
@@ -11,6 +11,7 @@ common-base = { path = "../base" }
 common-error = { path = "../error" }
 common-telemetry = { path = "../telemetry" }
 common-time = { path = "../time" }
+common-catalog = { path = "../catalog" }
 common-query = { path = "../query" }
 datatypes = { path = "../../datatypes" }
 snafu = { version = "0.7", features = ["backtraces"] }
