+++ src/common/grpc-expr/src/insert.rs
@@ -245,7 +245,10 @@ pub fn build_create_expr_from_insertion(
             }
         }
 
-        ensure!(timestamp_index != usize::MAX, MissingTimestampColumnSnafu);
+        ensure!(
+            timestamp_index != usize::MAX,
+            MissingTimestampColumnSnafu { msg: table_name }
+        );
         let timestamp_field_name = columns[timestamp_index].column_name.clone();
 
         let primary_keys = primary_key_indices
