+++ src/common/grpc-expr/src/lib.rs
@@ -18,7 +18,7 @@ mod column;
 pub mod error;
 mod insert;
 
-pub use alter::alter_expr_to_request;
+pub use alter::{alter_expr_to_request, create_expr_to_request, create_table_schema};
 pub use column::column_def_to_column_schema;
 pub use insert::{
     build_alter_table_request, build_create_expr_from_insertion, column_to_vector,
